+++ refactored/mgr/rdkStorageeMMC.cpp
@@ -98,7 +98,7 @@ eSTMGRReturns rStorageeMMC::populateDeviceDetails()
                     if((0 == strcasecmp(pDevType, "disk") && (0 == strcasecmp (pDevNode, m_devicePath.c_str())))) {
                         const char *pCapacity = udev_device_get_sysattr_value(pDevice, "size");
                         if(pCapacity) {
-                            m_capacity = (unsigned long) ((atol(pCapacity))*512);
+                            m_capacity = atoll(pCapacity) * 512;
                             STMGRLOG_INFO("[%s] m_capacity is : [%d]\n", __FUNCTION__, m_capacity);
                         }
                         const char* ro = udev_device_get_sysattr_value(pDevice, "ro");
@@ -125,7 +125,7 @@ eSTMGRReturns rStorageeMMC::populateDeviceDetails()
 
                             if (pCapacity)
                             {
-                                pPartitionPtr->m_capacity = (unsigned long) ((atol(pCapacity))*512);
+                                pPartitionPtr->m_capacity = atoll(pCapacity) * 512;
                                 STMGRLOG_INFO("[%s] Partition capacity is : [%lu]\n", __FUNCTION__, pPartitionPtr->m_capacity);
                             }
                             const char* ro = udev_device_get_sysattr_value(pDevice, "ro");
@@ -405,8 +405,8 @@ bool rStorageeMMC::get_eMMCPropertiesStatvfs()
                         return ret;
                     }
 
-                    unsigned long capacity  = vfs.f_blocks * vfs.f_frsize;
-                    unsigned long freeSpace = vfs.f_bavail * vfs.f_frsize;
+                    unsigned long long capacity  = vfs.f_blocks * vfs.f_frsize;
+                    unsigned long long freeSpace = vfs.f_bavail * vfs.f_frsize;
                     STMGRLOG_INFO ("Update the capacity n freespace as per STATVFS\n");
 
                     pObj->m_capacity = capacity;
